requirements.txt
fastapi
python-dotenv
uvicorn
httpx
PyJWT
passlib[bcrypt]
pydantic[email]
